commit
movement system made the bounciness occur with grrgrr, this cause the animations to not function due to the ground check required. Tried changing to a physics based movement system and ran into further problems. Using what I have now to build a completely different style of level that will work with the systems I have.